requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
gradio>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0

